Applied reactionary changes from Mesh to Generator
Removed the height and width arguments from the generate method of the
Generator interface. Added a compromise to the side length of meshes if
the specified length is larger than a dimension of the mesh that was
passed through, but kept the default sideLength as given for
reusability in the future.